+++ core/components/romanesco/elements/snippets/06_formulas/f_formblocks/fbloadassets.snippet.php
@@ -1,9 +1,15 @@
 <?php
+$assetsPathCSS = $modx->getOption('romanesco.custom_css_path', $scriptProperties, '');
 $assetsPathJS = $modx->getOption('romanesco.custom_js_path', $scriptProperties, '');
 $assetsPathVendor = $modx->getOption('romanesco.custom_vendor_path', $scriptProperties, '');
+$assetsPathDist = $modx->getOption('romanesco.semantic_dist_path', $scriptProperties, '');
 $uploadFile = $modx->getOption('uploadFile', $scriptProperties, 0);
 
+// Load Semantic UI form component separately
+$modx->regClientCSS($assetsPathDist . '/components/form.min.css');
+
 // Load FormBlocks JS in footer
+$modx->regClientScript($assetsPathDist . '/components/form.min.js');
 $modx->regClientScript($assetsPathJS . '/formblocks.js');
 
 // Load additional assets for file upload field, if present